+++ Data/ApplicationUser.cs
@@ -7,13 +7,13 @@ using Microsoft.AspNetCore.Identity;
 namespace WeTube.Data
 {
     // Add profile data for application users by adding properties to the ApplicationUser class
-    public class ApplicationUser : IdentityUser<int>
+    public class ApplicationUser : IdentityUser<long>
     {
         public virtual ICollection<Movie> Movies { get; set; }
         public virtual ICollection<Comment> Comments { get; set; }
     }
 
-    public class MyRole : IdentityRole<int>
+    public class MyRole : IdentityRole<long>
     {
         public MyRole() : base()
         {
